Revert "Changed file hierarchy"
